Improve defmulti tests
The iteration case was lacking a defmethod for the
:hydrangea dispatch value. It was also using a funky caught
error approach which was unnecessary. Instead, we implement
the default dispatch value case with a sentinal value.

The deftest also does not have to dump the values into
hashmaps, since it is not working with records.

I've also made the test clean up the namespace before and
after, and force a namespace reload, because this code
is (at the moment) not reloadable in the normal way.